Mudança no tipo do valor para a dto se adequar ao produto

+++ API_Banco/src/main/java/com/Gabriel/API_Banco/dto/CriarProdutoDTO.java
@@ -2,9 +2,11 @@ package com.Gabriel.API_Banco.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CriarProdutoDTO {
     private Long idLoja;
     private String nome;
-    private Double valor;
+    private BigDecimal valor;
 }
